CryptoAlg-574: Use a different alignment attribute
In versions of gcc older than 9.3, there is a check on `alignas` attribute on C++ files
that has a bug. [It was fixed in 9.3](https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89357#c3).
Here it's replaced by __attribute__((aligned(64))).
Note: alignas is used in other places of the code, but it seems that because
the value of the alignment is smaller than the number of entries in the arrays,
they don't cause errors.

Also, fixed a minor typo in `p256-armv8-asm.pl`.