CryptoAlg-591: Remove C code duplication from p256.c for ARMv8 nistz256 (#60)
### Issues:
CryptoAlg-591

### Description of changes: 
Remove C code duplication in p256.c for ARMv8 nistz256
Instead the C code in `p256-x86_64.[h,c]` is now used with bothx86_64 and AArch64 (ARMv8).
The code duplication was done in order to support gradual experimentation and development of ARMv8 P-256 optimization as in PRs 52, 56 and 57. In those PRs, the C functions that called Assembly functions were copied, few at a time, to `p256.c` to test the ARM assembly functions, one or more at a time.

### Call-outs:
In subsequent PRs, the files names will be changed, preferably after merging to main, so we have commits with the original name for ease of reference.

### Testing:
Tested on x86_64 (Macos) and ARMv8 (Graviton2).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.