C implementation of P-384 scalar multiplication functions
Implementation builds on top of the Fiat-crypto P-384 curve arithmetic.
The following functions are implemented:
  - ec_GFp_nistp384_point_mul,
  - ec_GFp_nistp384_point_mul_base,
  - ec_GFp_nistp384_point_mul_public,
alongside some additional helper functions. For the base point scalar
multiplication a precomputed table is used. The table is located in
in a new file (p384_table.h) which is generated by make_tables.go script.

Tests:
  - crypto_test passes including enabling,
  - `bssl speed` test results of ECDH and ECDSA (with NO_ASM=1):
               | OLD x86 | NEW x86 | Speedup |
    ECDH       |   962   |   1720  |   x1.8  |
    ECDSASign  |  1684   |   5314  |   x3.2  |
    ECDSAVerif |  1580   |   1885  |   x1.2  |

               | OLD GV2 | NEW GV2 | Speedup |
    ECDH       |   500   |    846  |   x1.7  |
    ECDSASign  |   875   |   2621  |   x3.0  |
    ECDSAVerif |   812   |    933  |   x1.1  |