Add manual override for NONOVERLAPPING_TAC
Now NONOVERLAPPING_TAC first looks for whatever it's trying to prove
in the goal's existing assumptions, either directly or with the
arguments swapped. This makes it easier to encourage things
semi-manually in cases where the default heuristics will not succeed,
typically on cases with the same base address and offsets involving
variables with a non-obvious relationship. In concert with this,
SIMPLE_ARITH_TAC now uses more restrictive filtering for some
efficiency gain.