Addressing CR comments
The main change is switching to scalar multiplication with window size
of 7 instead of 5. This is done to avoid a potential case of point
doubling in the multiplication function which would make the functions
non constant-time for certain scalars.

As a side-effect,the performance of ECDSA is improved on account of
using a larger precomputed table for the base point. Unfortunatelly,
ECDH is slightly slower now, but this will be analyzed and handled
in a separate piece of work.

- `bssl speed` test results of ECDH and ECDSA (with NO_ASM=1):
                 | OLD x86 | NEW x86 | Speedup |
      ECDH       |   962   |   1661  |   x1.7  |
      ECDSASign  |  1684   |   5835  |   x3.5  |
      ECDSAVerif |  1580   |   1939  |   x1.2  |

                 | OLD GV2 | NEW GV2 | Speedup |
      ECDH       |   500   |    808  |   x1.6  |
      ECDSASign  |   875   |   2866  |   x3.3  |
      ECDSAVerif |   812   |    964  |   x1.2  |