Add optimized 1024-bit squaring
The new function for 16-word (i.e. 1024-bit) squaring is called
"bignum_ksqr_16_32" with the "k" for Karatsuba, flagging the
requirement for the additional temporary buffer that is the last
argument. In fact the X86 version is a pure multiply array that
doesn't use the temporary buffer, but the API requires it for
consistency with the ARM version, which is indeed a Karatsuba
reduction. Because of their very different implementations the
proof is essentially trivial for X86 but more involved for ARM.