Add optimized 1024-bit multiplication
The new bignum_kmul_16_32 function, analogous to the earlier squaring
function bignum_ksqr_16_32, is a 16x16-word (i.e. 1024x1024-bit)
optimized multiplication, the ARM one a Karatsuba reduction using the
temporary buffer and the x86 one using the same API for consistency
even though it doesn't make use of the temporary buffer and is a
pure multiply-add array.