Switch x86 code from from NASM to GNU assembler
All x86/*/*.asm files have been turned into corresponding x86/*/*.S
files that can be assembled with the GNU assembler "as". This was a
customer request to minimize build dependencies, and anyway makes
sense from the point of view of uniformity because the Arm files
already use the GNU assembler.

The changes to each asm file are mostly routine and done by a script,
though in a few cases a macro was also slightly rewritten to be more
comprehensible making use of named arguments. On top of this there
are many additional cases where NASM would optimize loads of 32-bit
constants of the form

        mov     rax, 42

into a slightly shorter variant, relying on the zero-extension
guaranteed by the x86-64 semantics:

        mov     eax, 42

These cases have all been expressed explicitly for the GNU assembler,
since it doesn't make this optimization by default and arguably it's
stylistically nicer not to expect it.

The net effect is that all the x86/*/*.o files are exactly the same
as they were before, and hence the proof scripts x86/proofs/*.ml did
not need any change.