Add 256-bit and 384-bit big-endian conversions
The functions are called "bignum_bigendian_4" (4 words = 256 bits) and
"bignum_bigendian_6" (6 words = 384 bits), with additional aliases
using names "...frombytes..." and "...tobytes...", treating either
input or output as a byte array (exactly the same code, but with a
different type in the C prototype). Each such alias has its own
variant correctness theorem stated slightly differently.

Although they are not specific to the P-256 and P-384 primes (only
their native sizes) they are located in the respective "p256" and
"p384" subdirectories, since the main anticipated uses are in
connection with those primes.

For a little-endian platform, these always amount to byte reversal.
The ARM code has not been tested in a big-endian instantiation, and
since the current ARM model assumes little-endianness a priori,
neither has it been verified in a big-endian setting, but it is
supposed to work there. This is not relevant for x86, which is always
little-endian.