Remove separate almost-Montgomery functions for P-256 and P-384
The almost-Montgomery multiplications and squarings, where full
modular reduction is not guaranteed, do not seem to offer enough
(or in many cases any) performance advantage at the small sizes
of P-256 and P-384, in contrast to larger RSA-sized operations.
These separate functions have been completely deleted, but the
proofs of the "full" Montgomery operations for the NIST primes
(bignum_montmul_p256 etc.) have been augmented with variants
showing that they also work as almost-Montgomery functions. That
is, even without assumptions on the input ranges beyond those
implicit in the native sizes, the appropriate congruence is
guaranteed. It would perhaps be more elegant to put both of these
statements in the same proof, since the current change involves
some duplication of the proof scripts.