Slightly improve x86 P-256 and P-384 Montgomery operations
This commit makes various minor local improvements such as the use of
32-bit operations where possible (which generally reduces code size),
and generates some big constants avoiding the use of 64-bit immediate
moves. The Montgomery multiplies also do some of the top parts of
carry chains in a slightly more efficient way; this is the only case
where the actual dataflow changes, and even so the same proofs still
work modulo the instruction offsets and code size changes. Code size
is reduced for sure, and there seems to be some slight (maybe 1-3%?)
efficiency gain on both Xeon and EPYC. This commit also makes a couple
of tweaks to simulation in order to handle 32-bit ADC more nicely and
to trivialize the zero-extension of single-bit values "word(bitval b)".