Standardize numeric literals in ARM code to use # symbol
The code previously had a mix of cases where the "#" prefix was
used before numeric literals and cases where it wasn't. It makes
no difference to gcc or the GNU assembler (and hence this change
does not alter the object files or the proofs) but this makes
things more consistent and may help other tools processing the
ARM asm files.