Regularize .text and other directives
This makes sure each .S file has an explicit .text directive, and
also regularizes the positioning of these and the .globl ones (plus
the .intel_syntax ones for x86), collecting them at the beginning
just after the initial API comment banner. The only even slightly
non-trivial change is the removal of the global directive for
bignum_of_word_nontrivial in "x86/generic/bignum_of_word.S"
(this was not present for ARM, nor was it declared in the header).
There are no changes to code and hence no changes in the proofs.