Switch P-256 and P-384 code from assembler macros to C preprocessor
This is a superficial but syntactically messy change to avoid use of
the assembler .macro construct in the P-256 and P-384 functions, for
both ARM and x86 (so now macros are only used in the "fastmul"
functions), replaced by use of the C preprocessor. This is intended to
ease a FIPS build of the code, but does have the side-effect that
preprocessed code can now have multiple instructions per line
separated by semicolons. This is unproblematic for gcc or the GNU
assembler.

Took the chance to make one trivial but semantically meaningful
change, slightly tweaking x86/p384/bignum_tomont_p384.S in line with
earlier changes to the Montgomery multiplication to shorten the code a
bit and keep the identically-named "macros" the same.