Consistently use upper case for constants
Numeric constants from "#define" and assembler-level values
from ".set" are now always uppercase. This makes it easier to parse
the assembler files in a context-free way for scripting purposes.
This does not change the text section of any object file and hence
does not require proof changes. However the .set directives do
add a symbol to the symbol table, and those have changed in the
following from "i" to "I":

        arm/fastmul/bignum_kmul_16_32.o
        arm/fastmul/bignum_kmul_32_64.o
        arm/fastmul/bignum_ksqr_16_32.o
        arm/fastmul/bignum_ksqr_32_64.o
        arm/p256/bignum_mux_4.o
        arm/p384/bignum_mux_6.o
        arm/p521/bignum_double_p521.o
        x86/fastmul/bignum_kmul_32_64.o
        x86/fastmul/bignum_ksqr_16_32.o
        x86/fastmul/bignum_ksqr_32_64.o
        x86/p256/bignum_mux_4.o
        x86/p384/bignum_mux_6.o
        x86/p521/bignum_double_p521.o