Add AT&T syntax versions of x86 P-384 functions
The new files (x86_att/p384/*.S) are generated from the Intel syntax
originals (x86/p384/*.S) using a simple sed script. The generated
files are kept in this repo, but can be removed with "make clobber"
and regenerated with "make code" (in the new x86_att directory). This
translation process is not in general robust, only being intended to
handle this very restricted set of files. Even here, the translation
process runs a sanity check during generation to make sure each Intel
syntax original and its AT&T translation result in exactly the same
object file when assembled with the GNU assembler.