Add explicit alignment to ARM text sections
This uses ".balign 4" to force 4-byte (32-bit) alignment of the
instructions, as is assumed in the correctness theorems
("aligned_bytes_loaded ..."), though it seems likely to be the
default anyway as it's a basic requirement on ARM8.

Note that the equivalent ".align" directive is ".align 2" on ARM,
since the number is interpreted as an exponent 2^2 = 4. Since
this might be misleading at first sight and the meaning of
".align" differs between architectures and even output formats,
".balign" seemed preferable.

  https://ftp.gnu.org/old-gnu/Manuals/gas-2.9.1/html_chapter/as_7.html#SEC70

  https://developer.arm.com/documentation/100067/0611/armclang-Integrated-Assembler/Alignment-directives

s2n-bignum original commit: https://github.com/awslabs/s2n-bignum/commit/65f2a138dacc9a5f031ef28b9357fa8205b5845f