Remove spaces in function-like macro invocations
This involves transforming "f (arg1,...,argn)" -> "f(arg1,...,argn)"
on invocations of each macro f, since there exist preprocessors that
don't expand the macro if the whitespace is there. Arguably their
behavior is buggy, at least in the context of actual C compilation,
but removing the space seems if anything more readable and natural.
(Note that this is a different question from spaces in the
*definition* of a macro, where of course they would make it an
object-like macro.)