Add AT&T syntax versions of all other x86 functions
Now there is a full set of AT&T translations of the x86 code. The sed
script doing the translation is now a bit cleaner and has some
additional lines to handle the macros used in the code. It was easier
to slightly tweak the original x86/fastmul/bignum_kmul_32_64.S
rather than complicate the sed script, but conversely the mild
uglification made earlier to x86/generic/bignum_modinv.S for ease of
processing is now reversed. In both cases, the object files and
hence proofs are of course unchanged.