Correctly check Jitter read_entropy return value (#248)
Documentation for Jitter read_entropy functions states that the functions return either a negative integer which signifies an error, or the number of generated bytes of entropy. The implementation of the functions however checks if it the requested and the generated number of bytes is the same and returns an error if it's not. Nevertheless, to be on the safe side we now check that Jitter function returns the expected value.

Co-authored-by: Dusan Kostic <dkostic@amazon.com>