Always use SHA-256 for FIPS integrity check
FIPS 140-3 requires a Cryptographic Algorithm Self-Test (CAST)
for the approved integrity technique used for the integrity test.
The CAST must be performed to verify the integrity technique itself
and it must happen before the module performs the integrity test.

Since HMAC-SHA2-512 is not tested in our power-on self tests,
and HMAC-SHA2-256 is we change the code to always use SHA-256
based HMAC for integrity check. Also, we reorder the sequence
of function calls to perform the CAST of HMAC-SHA2-256 before
the integrity check.