Update rand.c
### Issues:
Addresses CryptoAlg-864

### Description of changes: 
When in FIPS mode we use the CPU Jitter entropy source to seed our DRBG. This entropy source is very slow and can incur a cost anywhere between 10-60ms depending on configuration and CPU.  Increasing to 2^24 will amortize the penalty over more requests.  The value was selected by reviewing OpenSSL and SP 800-90B for the max reseed interval (2^48).

### Call-outs:
N/A

### Testing:

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.