Use pq-crypto
Move the pq-crypto folder under crypto and use (the original SIKE submission) files for SIKE. The crypto/sike/ dir is removed from the crypto/CMakeList.txt to avoid conflicts. The SIKE functions are not hand-coded renamed in the hpke.c file, rather are renamed using directives. The return value of the SIKE function (s2n and pq-crypto integration designs) differ, we have changes the check value from if(! sike_keypair(.., ..)), to if(sike_keypair(.., ..)<0). We keep both dirs, to use s2n reverse the changes or use previous commit