Benchmark all
Wrap the functions setup_sender, setup_recipient, seal, open with a cpucycles() function to get the performance. The time required per function is expressed in clock cycles x10^6 and in percentage compared to the entire protocol execution. The total time of the protocol does not include the allocation of the arrays for the ciphertext and plaintext since it becomes too expensive with the increase of the arrays. To run the tests change the HPKE_VERSION_PQ directive to 0 for x25519, 1 for SIKE and 2 for x25519+SIKE, which is place inside the inlcude/openssl/hpke.h file, install the library and run the tests with respectively ninja install && ./crypto/crypto_test --gtest_filter=HPKETest.{x25519, SIKE, Hybrid} >> file_with_results.txt