Add word-level reciprocal approximation
If input "a" is a 64-bit unsigned word with MSB set (i.e. a >= 2^63)
the value x = word_recip(a), when augmented by an implicit leading 1
bit to give x' = 2^64 + x, satisfies x' = ceil(2^128/a) - 1, which in
all cases except a = 2^63 exactly is the same as x' = floor(2^128/a).
In keeping with the constant-time style, this is done avoiding any
use of the machine division instruction, using an initial polynomial
approximation followed by Newton iteration.