Clean up a few common patterns in proofs
This factors out as BIGNUM_OF_WORDLIST_DIV_CONV the evaluation of
`bignum_of_wordlist [...; ... ; ... ] DIV 2 EXP ...`, and as
SHARPEN_INEQ_TAC the sharpening using discreteness of an ostensibly
real inequality goal where both sides can be deduced to be integers.
It also makes more use of the improved BOUNDER_TAC, which is still
further improved.