Add UBFM to ARM model and make other improvements
The main enhancement is to add the general case of the UBFM (unsigned
bitfield move) instruction to the formal ARM model. The basic left and
right shifts are now done properly as special cases: although there
are still pseudo-instructions for them at the shallow embedding level,
the decoding function gives the UBFM form that is proved equivalent
later in the simulation process.

This update also fills out and regularizes the existing suite of basic
operations in a natural way by adding ASRV, BIC, BICS and RORV, as
well as calling out a few additional aliases. Finally, it fixes an
error in decoding of a couple of instruction classes where the
decision over the interpretation of register 31 as the zero register
or stack pointer was insufficiently nuanced (basically, it depends
in these cases on whether it is an "S"-suffixed flag-setting form).