Removed jitterentropy static library file from install
The jitterentropy static library file was previously getting installed
when running make install. Since this is just a file we link against and
not a shared library file or a static library file that is part of our
release we don't need to include this upon running make install.

This change removes the install logic from the jitterentropy CMake file
and modifies the crypto CMake file to do private linking against
jitterentropy. In order to use private linking I needed to explicitly
list all the other linking types as public.