Removed jitterentropy static library file from shared install (#306)
* Removed jitterentropy static library file from install when compiling a shared library

The jitterentropy static library file was previously getting installed
when running make install. Since this is just a file we link against and
not a shared library file, when we build a shared library we don't need it.

This change removes the install logic from the jitterentropy CMake file
and modifies the crypto CMake file to do private linking against
jitterentropy. In order to use private linking I needed to explicitly
list all the other linking types as public.