Test the coordinate range validation in `EC_KEY_check_fips` (#322)
This commit tests the point validation checks added inside `EC_KEY_check_fips` in b80c18f.
The points used in the test have a small x-coordinate. If corresponding y coordinates exist both points are added to the test vectors (y_0 \in {0,1}).
As explained in ec_test.cc, the checks are only exercised in the case of P-224 when its EC_METHOD members use regular representation of field elements from `simple.c` and not Montgomery representation, which is the case in `EC_GFp_nistp224_method`.