Init Service Indicator counter in `FIPS_service_indicator_update_state` (#344)
FIPS 140-3 requires that the module should provide the service indicator for approved services irrespective of whether the user queries it or not. Hence, we cannot solely depend on the external call  |FIPS_service_indicator_before_call| to initialize it.

This change adds a potential initialization in FIPS_service_indicator_update_state() which is an internal call in all approved services. The effect of this change is that, after the power-on self-test, the counter would already be incremented more than once.