Add simple benchmarking program
The benchmarking procedure times each s2n-bignum function in turn
(excluding those on x86 where the machine doesn't support necessary
BMI and ADX extensions). In the case of functions accepting inputs
of arbitrary sizes, at least one and sometimes (where it seems more
interesting) several sizes are exercised.

For each function, pseudo-random inputs with all integer bit densities
from 0 to 64 are considered. As well as average values, the coefficient
of variation and the (Pearson) correlation coefficient with bit
density are shown for each function. Since these functions are all
intended to be constant-time (i.e. independent of data other than the
size parameter if applicable), the coefficient of variation and
correlation coefficient ought both to be small.