Enabling the shared FIPS build on macOS (#350)
This commit enables the shared FIPS build on macOS.

To properly build the FIPS module on macOS with Apple’s toolchain
we have to use a different method (than on Linux) for marking
the start and end of the module for the FIPS integrity check.

On Linux, we use linker scripts to make sure that all the code is
in the text section of the module between |BORINGSSL_bcm_text_start|
and |BORINGSSL_bcm_text_end| markers and all the read-only data is
in the data section between |BORINGSSL_bcm_rodata_start| and
|BORINGSSL_bcm_rodata_end| markers.

Apple's linker (ld) doesn't support linker scripts so we implement the
"markers" similarly to OpenSSL. We add a dummy file |fips_macos_support.c|
that we compile twice:
  - Once with |AWSLC_FIPS_MACOS_START| flag. This will produce
    |fips_macos_start.o| object file that contains |BORINGSSL_bcm_text_start|
    and |BORINGSSL_bcm_rodata_start| in the text and data sections.
  - The second time with |AWSLC_FIPS_MACOS_END| flag. This will produce
    |fips_macos_end.o| object file that contains |BORINGSSL_bcm_text_end|
    and |BORINGSSL_bcm_rodata_end| in the text and data sections.
Once we have |fips_macos_start.o|, |fips_macos_end.o| and the object file
of the actual module |bcm.o|, we link the final FIPS module by specifying
the files in the following order when linking:
    ld fips_macos_start.o bcm.o fips_macos_end.o ...
On macOS this will ensure that all the code of the module is in the __text
section and all the read-only data is in the __const section between the
corresponding markers.

The above allows us to properly compute the hash of the module's text and
data sections and perform the FIPS integrity check. As it's done on Linux,
after we build the module and |libcrypto.so|, the script |inject_hash.go|
parses the library (|libcrypto.dylib| when compiling on macOS), computes
the hash of the module and stores it in a previously allocated place for it.
To do this on macOS we updated |inject_hash.go| to add code that can parse
Apple specific Mach object file format (Mach-O).