Optimize modular inverse inner loop
The new version restructures the inner (word-level) loop. The net
result is exactly the same as before, but the computation uses fewer
instructions and/or has better dependent latency, especially on ARM.
For relatively small size, e.g. those used in common elliptic curves,
the speedup is significant. At larger RSA-type sizes the
multiplications dominate anyway so the overall benefit is relatively
more modest. The same changes are applied to the coprimality
predicate, which uses the same basic inner loop.