Improve treatment of short operations in x86 model
This enables 16-bit operands in the decoder by interpreting the
operand size override prefix appropriately. This is currently only
enabled for a subset of instructions, but could be done more widely
by changing more instances of "if has_pfxs pfxs then NONE else ..."
to the new "if has_unhandled_pfxs pfxs then NONE else ...". The
existing treatment of 8-bit operands in the model is also improved,
introducing names for the register portions SPL, BPL, SIL and DIL
and correctly distinguishing the decoder mapping to those or to
AH, CH, DL and BL according to the use of the REX prefix. The
simulation machinery also more systematically expands out reads
and writes to short registers in terms of the full register,
previously just done for CL in an ad-hoc way (because of its use
in variable shifts).