Integrate s2n-bignum alternative functions for P-384 (#386)
On x86_64 platforms s2n-bignum uses bmi2 and adx instruction sets
for some of the P-384 functions. These instructions are not supported
by every x86 CPU. Previously, we checked in runtime if bmi2 and adx
are supported and if not we fallback to Fiat's implementation. This
solution had a drawback that we had to always compile both
s2n-bignum and Fiat code.

s2n-bignum implemented "alternative" implementations of functions
that don't require any special instructions. In this commit we integrate
those changes and instead of falling-back to Fiat's code we fall back
to s2n-bignum alternative implementations. The change has two
benefits:
  - s2n-bignum and Fiat are decoupled so we compile only one.
  - Performance improvement on x86 CPUs that don't support bmi2
     and adx instructions is x2.11, x1.58, x2.03 for ECDH, ECDSA sign,
     ECDSA verify, respectively.

The benchmark for a CPU without bmi2 and adx support, with
bssl/speed -filter “P-384”, before and after:
ECDH: 1594 —> 3372 (ops/s)
ECDSA Sign: 6332 —> 10000 (ops/s)
ECDSA Verify: 1981 —> 4019 (ops/s)