Add mangled names for globals to each function for Mac OS
This includes name-mangled forms starting with underscores for all the
globals, in addition to the existing unmangled ones. At the cost of
adding redundant names, this simplifies building on Mac OS. Previously
this was done at assemble-time by an additional step in the pipeline
of transformations on the source code, which has been removed now.