Tweak quotient multiplication on x86 curve25519 field operations
Where there is a product 19 * q in which we know q <= 77, it's
very safe to assume this fits in 64 bits, and hence we can just
use imul and not generate the upper part of the product. This
seems to give a noticeable gain in speed for three of the functions.