Support old assemblers that don't know about AVX (#398)
We assume that the assembler that is going to be used to compile the assembly code in the library is modern enough that it knows about AVX. However, this assumption is not always correct for our internal systems. With this patch we add logic to handle the case of ancient compilers that don't support AVX. For this purpose we define a CMake flag called MY_ASSEMBLER_IS_TOO_OLD_FOR_AVX that can be set to TRUE in case the assembler doesn't understand AVX instructions. In perl-assembly files we check if MY_ASSEMBLER_IS_TOO_OLD_FOR_AVX flag is specified and don't generate the AVX code in case it is. Regarding pure assembly files that have AVX implementations -- we have to remove them from compilation all together.

Note that we also simplified and unified how our Perl-assembly scripts handle command line arguments. Previously, some scripts were expecting the flavour and output arguments to be the first two arguments (if given), while other scripts assumed flavour to be the first one and output to be the last argument. Now, all the scripts require at least two arguments, the first one being the flavour, the second one being the output file path, and optionally, additional flags starting from the third argument.