P-521 implementation with s2n-bignum field arithmetic (#412)
This commit adds the implementation of P-521 curve that uses s2n-bignum
field arithmetic. The s2n-bignum library offers assembly implementation
optimized for x86_64 and aarch64 architectures. The integration is done
in the same way as for P-384. Basically, s2n-bignum field arithmetic
functions are designed as a drop-in replacement for their Fiat-crypto
counterparts. The one big difference between Fiat and s2n-bignum in
this case is that they use a different representation of field elements.
Because of this we need to generate different table of pre-computed
multiples of the base point for s2n-bignum. The rest of the integration
is a straightforward replacement of Fiat with s2n-bignum functions.

Finally, the performance improvements of this change are:

Intel(R) Xeon(R) Platinum 8124M CPU (x86_64)
| Operation           | Before(ops/sec) | After(ops/sec) | Improvement |
| ECDH P-521          |       1610      |      2865      |     1.8x    |
| ECDSA P-521 signing |       4300      |      5693      |     1.3x    |
| ECDSA P-521 verify  |       1767      |      2781      |     1.6x    |

Graviton 2 (ARMv8, aarch64)
| Operation           | Before(ops/sec) | After(ops/sec) | Improvement |
| ECDH P-521          |        532      |       802      |     1.5x    |
| ECDSA P-521 signing |       1460      |      1748      |     1.2x    |
| ECDSA P-521 verify  |        575      |       825      |     1.4x    |