OpenSSL CVE-2022-0778 Fix possible infinite loop in BN_mod_sqrt()
This commit fixes the issue released as OpenSSL CVE-2022-0778 that
affects AWS-LC as well. A bug in BN_mod_sqrt() can cause the function
to enter an infinite loop. The issue is now fixed and two test
cases are added to verify that the function returns a failure instead
of hanging.