P-384 and P-521 scalar multiplication with smaller tables (#418)
P-384 and P-521 scalar multiplication with smaller tables 

This commit changes the size of the precomputed tables used in scalar
multiplication fuctions of P-384 and P-521 implementations.
Our experiments show that we can have approximately the same performance
with much smaller tables. Previously, we were using tables of size:
  - 86KB and 175KB for the precomputed table of base point multiples
    in P-384 and P-521 case, respectively. This table is used in
    mul_base and mul_public functions.
  - 9KB and 14KB for the table computed on-the-fly in mul function
    in P-384 and P-521 case, respectively.
Note that the tables in the first case are precomputed and are part
of the binary (libcrypto.so), while the tables in the second case
are computed in runtime and don't affect the binary size.

With this commit we change the tables and obtain the following savings:
  - the precomputed table: from 86KB to 30KB and from 175KB to 62KB,
  - the on-the-fly table: from 9KB to 2KB and from 14KB to 4KB,
in P-384 and P-521 case, respectively. The performance comparison
between the new and the old tables is shown in the tables below
(the performance is measure in number of operations per second by the
bssl speed tool).

Intel(R) Xeon(R) Platinum 8124M CPU (x86_64)
  | Operation           |  Old (86KB+9KB)   |  New (30KB+2KB)   |
  | P-384 ECDH          |       3747        |       4092        |
  | P-384 ECDSA sign    |      10651        |      10510        |
  | P-384 ECDSA verify  |       4457        |       4324        |
   -------------------------------------------------------------
  |                     | Old (175KB+14KB)  |  New (62KB+4KB)   |
  | P-521 ECDH          |       2857        |       3088        |
  | P-521 ECDSA sign    |       5673        |       5679        |
  | P-521 ECDSA verify  |       2782        |       2698        |

Graviton 2 (ARMv8, aarch64)
  | Operation           |  Old (86KB+9KB)   |  New (30KB+2KB)   |
  | P-384 ECDH          |       1787        |       2009        |
  | P-384 ECDSA sign    |       4447        |       4480        |
  | P-384 ECDSA verify  |       2085        |       2085        |
   -------------------------------------------------------------
  |                     | Old (175KB+14KB)  |  New (62KB+4KB)   |
  | P-521 ECDH          |        799        |        877        |
  | P-521 ECDSA sign    |       1748        |       1744        |
  | P-521 ECDSA verify  |        827        |        810        |