Add Android test harness for AWS-LC (#389)
* AWS-LC wants to ensure that our tests and build work correctly 
on Android. Insipred from the CRT Team's 
`AWSCRTAndroidTestRunner`, this mini Android test harness is
 intended to be used within our Android CI with AWS Device Farm.
 The tests will include crypto_test, urandom_test, decrepit_test
and ssl_test, referenced from regular test dimensions. 
Using this test harness with AWS-LC will also help ensure that 
builds cross compiles successfully with the Android toolchains,
mitigating past issues like this: 
https://github.com/awslabs/aws-lc/issues/192.

* ANDROIDTESTRUNNER is a new flag that I plan to add in AWS-LC 
that compiles our test executables from regular executables to 
so executables when ANDROIDTESTRUNNER is on. Only Shared 
Object files can be automatically packaged into an Android app 
with the right execution permissions, so test executable outputs 
need to be outputted as an so for AWSLCAndroidTestRunner to be
 able to execute the tests within the app dimension.
The actual ANDROIDTESTRUNNER change in AWS-LC is not in this 
PR, and this is just an early call out to the change I intend to make.