Add Windows ABI wrappers
The x86 ABI on Windows differs from the "standard" one used on Linux,
Mac OS etc. For our purposes the relevant differences are that RDI and
RSI are callee-saved (need to be preserved if modified), and that the
input arguments are in different places. Integer return values are
still in RAX. Here are the conventions for the first 6 integer
arguments (the maximum that s2n-bignum functions rely on so far):

   Argument number       Standard        Windows

          1                RDI            RCX
          2                RSI            RDX
          3                RDX             R8
          4                RCX             R9
          5                 R8           [RSP+40]
          6                 R9           [RSP+48]

Here we add a wrapper round each s2n-bignum function that optionally
(controlled by the WINDOWS_ABI variable in preprocessing) makes the
function compatible with the Windows ABI. This is done in a simplistic
way simply by pushing the extra callee-saved registers (RSI and RDI),
shuffling arguments (only the ones actually used by that function),
performing the same computation in exactly the original way, and
finally popping back the extra registers.

The Makefile currently selects this WINDOWS_ABI mode when `uname -s`
gives "CYGWIN_NT-10.0", as on the main test platform, but this can be
configured.