Add Android cross-compilation CI dimension with `AWSLCAndroidTestRunner`  (#448)
This change adds a new codebuild batch for android builds 
`Static`, `Shared`, `Debug`, `Release`, and `FIPS`. It also
adds the `ANDROIDTESTRUNNER` flag which compiles our test
executables into `so` executables when it's on. Only Shared 
Object files can be automatically packaged into an Android app
with the right execution permissions, so test executable
outputs need to be outputted as an `so` for our
`AWSLCAndroidTestRunner` app to be able to execute the tests
within the app dimension.
The yml files under `tests/ci/codebuild/android/` were split
into 5 dimensions for each build, so we can parallelize the
Device Farm runs once the Device Farm functionality is turned
on.